pyproject.toml
[tool.poetry]
name = "miskatonic"
version = "0.1.0"
description = ""
authors = ["Dan Elkayam <dan.elkayam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.11"
django = "^4.1.7"
django-extensions = "^3.2.1"
psycopg2-binary = "^2.9.5"
dj-database-url = "^1.2.0"
pre-commit = "^3.1.0"
requests = "^2.31.0"
pydantic = "^1.10.5"
pydantic-django = "^0.1.1"
django-ninja = "^0.21.0"
beautifulsoup4 = "^4.12.0"
numpy = "^1.25.0"
jupyter = "^1.0.0"
pandas = "^2.0.3"
django-pandas = "^0.6.6"
torch = "^2.0.1"
tqdm = "^4.65.0"
django-postgres-extra = "^2.0.8"
pytorch-nlp = "^0.5.0"
matplotlib = "^3.7.2"
scikit-learn = "^1.3.0"
ray = { extras = ["air"], version = "^2.5.1" }
memray = "^1.8.1"
gputil = "^1.4.0"
gensim = "^4.3.1"
nltk = "^3.8.1"
lightning = "^2.0.6"
tensorboard = "^2.13.0"
comet-ml = "^3.33.8"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
